==> 00_inventory_analysis_&_curve_generation/03_yield_curve_convert.ipynb <==
Convert TIPSY output to ws3 input yield curve table 

#Create non_fert wood allocation curve and save as base curve_points_table_0

import numpy as np
import pandas as pd

fert=0
#Create fertilized yield curve and add fertilized curve id into au_table 
dft = pd.read_csv('./data/tipsy_curves_tsa24_%s.csv' % fert)

# df_curve_table_original = pd.read_csv('./data/curve_points_table.csv')
# dfcto = df_curve_table_original
curve_points_table_data = {'curve_id':[],
                           'age':[],
                           'volume':[]}

cruve_table_data = {
                    'curve_id': [],
                    'curve_type': ['managed']
                    }



dft_au = dft['AU'].unique()
dft_au_tsa = dft_au + 2400000
dft['fert_curve_id'] = dft['AU'] + 2400000 + (fert * 10)

# dfcto['age'] = dfcto['x']
# dfcto['volume'] = dfcto['y']

curve_points_table_data['curve_id'] = dft['fert_curve_id']
curve_points_table_data['age'] = dft['Age']
curve_points_table_data['volume'] = dft['Vol_gross']
curve_points_table_data['ps124'] = dft['2+_2*4'] #s124 represent grade two plus lumber(grade s, grade 1 and grade 2) and it's 2 by 4 (2 inch by 4 inch)
curve_points_table_data['ps126'] = dft['2+_2*6']
curve_points_table_data['ps128'] = dft['2+_2*8']
curve_points_table_data['ps1210'] = dft['2+_2*10']
curve_points_table_data['p34'] = dft['3_2*4'] #34 represent grade three lumber and it's 2 by 4
curve_points_table_data['p36'] = dft['3_2*6']
curve_points_table_data['p38'] = dft['3_2*8']
curve_points_table_data['p310'] = dft['3_2*10']
curve_points_table_data['p44'] = dft['4_2*4']
curve_points_table_data['p46'] = dft['4_2*6']
curve_points_table_data['p48'] = dft['4_2*8']
curve_points_table_data['p410'] = dft['4_2*10']
curve_points_table_data['pbark'] = dft['Bark']
curve_points_table_data['pchips'] = dft['Chips']
curve_points_table_data['psawdust'] = dft['Sawdust']
curve_points_table_data['pshaving'] = dft['Shavings']
curve_points_table_data['ptrim'] = dft['Trims']

curve_points_table = pd.DataFrame(curve_points_table_data)

# curve_points_table_updated = pd.concat([curve_points_table, df_curve_table_original])
# curve_points_table_updated = curve_points_table_updated.drop(columns=['x'], axis=1)
# curve_points_table_updated = curve_points_table_updated.drop(columns=['y'], axis=1)
# curve_points_table_updated = curve_points_table_updated.drop(columns=['Unnamed: 0'], axis=1)
curve_points_table.to_csv('./data/thesis_input/curve_points_table_0.csv', index=False)




import numpy as np
import pandas as pd

fert=0

#Create fertilized yield curve and add fertilized curve id into au_table 
dft = pd.read_csv('./data/tipsy_curves_tsa24_%i.csv' %fert)
dfa = pd.read_csv('./data/au_table.csv')
df_curve_table_original = pd.read_csv('./data/curve_points_table_0.csv')
dfcto = df_curve_table_original
curve_points_table_data = {'curve_id':[],
                           'age':[],
                           'volume':[]}

cruve_table_data = {
                    'curve_id': [],
                    'curve_type': ['managed']
                    }



dft_au = dft['AU'].unique()
dft_au_tsa = dft_au + 2400000
dft['fert_curve_id'] = dft['AU'] + 2400000 + (fert * 10)


curve_points_table_data['curve_id'] = dft['fert_curve_id']
curve_points_table_data['age'] = dft['Age']
curve_points_table_data['volume'] = dft['Vol_gross']
curve_points_table_data['ps124'] = dft['2+_2*4'] #s124 represent grade two plus lumber(grade s, grade 1 and grade 2) and it's 2 by 4 (2 inch by 4 inch)
curve_points_table_data['ps126'] = dft['2+_2*6']
curve_points_table_data['ps128'] = dft['2+_2*8']
curve_points_table_data['ps1210'] = dft['2+_2*10']
curve_points_table_data['p34'] = dft['3_2*4'] #34 represent grade three lumber and it's 2 by 4
curve_points_table_data['p36'] = dft['3_2*6']
curve_points_table_data['p38'] = dft['3_2*8']
curve_points_table_data['p310'] = dft['3_2*10']
curve_points_table_data['p44'] = dft['4_2*4']
curve_points_table_data['p46'] = dft['4_2*6']
curve_points_table_data['p48'] = dft['4_2*8']
curve_points_table_data['p410'] = dft['4_2*10']
curve_points_table_data['pbark'] = dft['Bark']
curve_points_table_data['pchips'] = dft['Chips']
curve_points_table_data['psawdust'] = dft['Sawdust']
curve_points_table_data['pshaving'] = dft['Shavings']
curve_points_table_data['ptrim'] = dft['Trims']

curve_points_table = pd.DataFrame(curve_points_table_data)


# List of columns to be summed
columns_to_sum = ['ps124', 'ps126', 'ps128', 'ps1210', 'p34', 'p36', 'p38', 'p310', 'p44', 'p46', 'p48', 'p410']

# Using DataFrame.assign() to add new column 'lumber' which is the sum of the columns in columns_to_sum
curve_points_table = curve_points_table.assign(lumber=curve_points_table[columns_to_sum].sum(axis=1))


curve_points_table_updated = pd.concat([curve_points_table, df_curve_table_original])
curve_points_table_updated.to_csv('./data/thesis_input/curve_points_table_updated_%s.csv' % fert, index=False)

#Create au table based on existed au table
#Cannot modify au_table, create dfa_fert for specific fert scenario
dfa_fert = dfa.copy()
dfa_fert['fert_curve_id'] = 0
for index,row in dfa_fert.iterrows():
    if row['managed_curve_id'] != row['unmanaged_curve_id']:
        dfa_fert.loc[index, 'fert_curve_id'] = row['managed_curve_id'] + (fert * 10)
    else:
        dfa_fert.loc[index, 'fert_curve_id'] = row['managed_curve_id'] 

dfa_fert.to_csv('./data/thesis_input/au_table_%s.csv' % fert, index=False)



dfa_fert

curve table no need to be updated every time

# cruve_table_data = {
#                     'curve_id': [],
#                     'curve_type': ['managed']
#                     }

# unmanaged_curve_id = dfa['unmanaged_curve_id'].unique()
# managed_curve_id = dfa['managed_curve_id'].unique()
# managed_curve_id = np.append(dfa['managed_curve_id'].unique(), dfa['fert_curve_id'].unique())
# # managed_curve_id.extend(dfa['fert_curve_id'].unique())
# managed_curve_id = np.unique(managed_curve_id)
# managed_curve_id
# curve_id_data = np.append(unmanaged_curve_id,managed_curve_id)
# curve_id_data = np.unique(curve_id_data)
# curve_id_data

# df_curve_table_data = pd.DataFrame.from_dict(curve_table_data, orient="columns")
# curve_table_data['curve_id'] = curve_id_data
# # df_curve_table_data = pd.DataFrame.from_dict(curve_table_data, orient="columns")
# for index, row in df_curve_table_data.iterrows():
#     if row['curve_id'] in unmanaged_curve_id:
#         df_curve_table_data.loc[index, 'curve_type'] = 'unmanaged'
#     else:
#         continue

# df_curve_table_data.to_csv('./data/curve_table_%s.csv' % fert, index=False)
